First set of Core Classes changes! w00t :P
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@823 3db68df2-c116-0410-a063-a993310a9797

+++ source/simulation/Entity.cpp
@@ -158,7 +158,7 @@ void CEntity::snapToGround()
 	m_graphics_position.Y = g_Terrain.getExactGroundLevel( m_graphics_position.X, m_graphics_position.Z );
 }
 
-void CEntity::update( float timestep )
+void CEntity::update( size_t timestep )
 {
 	m_position_previous = m_position;
 	m_orientation_previous = m_orientation;
